+++ dataprep-backend-service/src/main/resources/org/talend/dataprep/messages.properties
@@ -28,6 +28,8 @@ import.hdfs.title=Add HDFS dataset
 
 import.job.label=From Talend Job
 import.job.title=Add Talend Job dataset
+import.tcomp/FullExampleDatastore.label=From TCOMP example
+import.tcomp/FullExampleDatastore.title=Add a TCOMP dataset !
 
 parameter.name.desc=Name
 parameter.name.label=Dataset name
