+++ dataprep-backend-service/src/test/java/org/talend/dataprep/api/dataset/json/DataSetJSONTest.java
@@ -32,10 +32,10 @@ import org.talend.dataprep.api.dataset.ColumnMetadata;
 import org.talend.dataprep.api.dataset.DataSet;
 import org.talend.dataprep.api.dataset.DataSetContent;
 import org.talend.dataprep.api.dataset.DataSetMetadata;
+import org.talend.dataprep.api.dataset.location.LocalStoreLocation;
 import org.talend.dataprep.api.dataset.row.DataSetRow;
 import org.talend.dataprep.api.type.Type;
 import org.talend.dataprep.dataset.DataSetMetadataBuilder;
-import org.talend.dataprep.dataset.HttpLocation;
 import org.talend.dataprep.exception.TDPException;
 import org.talend.dataprep.exception.error.CommonErrorCodes;
 import org.talend.dataprep.schema.csv.CSVFormatFamily;
@@ -156,8 +156,7 @@ public class DataSetJSONTest extends ServiceBaseTests {
         content.setMediaType("text/csv");
         metadata.getLifecycle().qualityAnalyzed(true);
         metadata.getLifecycle().schemaAnalyzed(true);
-        HttpLocation location = new HttpLocation();
-        location.setUrl("http://estcequecestbientotleweekend.fr");
+        LocalStoreLocation location = new LocalStoreLocation();
         metadata.setLocation(location);
         StringWriter writer = new StringWriter();
         DataSet dataSet = new DataSet();
