+++ dataprep-dataset/src/main/java/org/talend/dataprep/api/dataset/location/AbstractUrlLocation.java
