+++ dataprep-dataset/src/main/java/org/talend/dataprep/dataset/service/BaseDataSetService.java
@@ -18,5 +18,4 @@ public abstract class BaseDataSetService {
             throw new TDPException(DataSetErrorCodes.UNABLE_TO_SERVE_DATASET_CONTENT, context);
         }
     }
-
 }
