+++ dataprep-dataset/src/test/java/org/talend/dataprep/api/dataset/location/HttpLocationTest.java
@@ -18,7 +18,6 @@ import static org.junit.Assert.assertNull;
 
 import org.apache.commons.httpclient.URIException;
 import org.junit.Test;
-import org.talend.dataprep.dataset.HttpLocation;
 
 /**
  * Unit test for the HttpLocation
