+++ dataprep-dataset/src/test/java/org/talend/dataprep/dataset/store/content/http/HttpContentStoreTest.java
@@ -19,8 +19,8 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 
 import org.junit.Test;
+import org.talend.dataprep.api.dataset.location.HttpLocation;
 import org.talend.dataprep.dataset.DataSetBaseTest;
-import org.talend.dataprep.dataset.HttpLocation;
 
 /**
  * Unit test for the remote http datasets.
