+++ dataprep-dataset/src/test/resources/org/talend/dataprep/api/dataset/location/locator/hdfs_location_ok.json
