+++ dataprep-dataset/src/test/resources/org/talend/dataprep/api/dataset/location/locator/http_location_ok.json
